Add files via upload
Here is the homework for the second part, I tried to do it through subquery, CTE as well as to combine joins. I didn't get solutions for everything because I don't have any years after 2006 in the database, so I'm not sure if I'm getting adequate results
I divided the sql files so that the domestic number 1:1,1.1,1.2,1.3,1.4
homework no 2 : 2.1,2.1
and homework no 3: 3,3.1
I hope I didn't make a lot of mistakes and if it's not difficult for you to write me the mistakes so that I know for later. thank you in advance for your patience